back/requirements.txt
fastapi
pydantic
pymongo
uvicorn
pytest
requests
python-dotenv